Enable update in further methods and in Moveable. Disable minimap from TileMap in main, because with the FPS boost such things cannot be done. It seems the FPS Boost is only possible with a static or hardly dynamic scene. I will discontinue the experiment for an unknown time.

+++ Window/Window.d
@@ -294,7 +294,7 @@ final:
 		ubyte[] tmpLine = Array!ubyte(&ta)[lineWidth];
 		debug writeln("Screenshot alloc: ", tmpLine.length, "::", lineWidth);
 		
-		for (uint i = 0; i < this.height / 2; ++i) {
+		for (ushort i = 0; i < this.height / 2; ++i) {
 			const uint tmpIdx1 = i * lineWidth;
 			const uint tmpIdx2 = (i + 1) * lineWidth;
